Added search functionality, User Authentication, non-auth page restriction, in middle of user registration. Sorry, was in the habit of updating 2019 project between every video ebcasue those were 20 min segments & this course is a single 7 hr jawn, but I should get back into the habit of more frequent, compartmentalized updates. I know I'm not branching off of git and using it's features to the fullest but I will hopefully be on analogous projects in the future, so I should build that habit early. Also I gotta add more terse messages

+++ deckchat/base/views.py
@@ -1,7 +1,12 @@
 from multiprocessing import context
 import re
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q 
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm 
 
@@ -10,6 +15,47 @@ from .forms import RoomForm
 #    {'id':2, 'name':'Jeskai Murktide'},
 #    {'id':3, 'name':'Naya Depths'},
 #]
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        try:
+            password = User.objects.get(password=password)
+        except:
+            messages.error(request, 'Incorrect password')
+            
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'username or password does not exist')
+
+       
+    context = {'page': page}
+    return(render(request, 'base/login_register.html', context))
+
+
+def logoutUser(request):
+    logout(request)
+    return(redirect('home'))
+
+def registerPage(request):
+    page = 'register'
+    return render(request, 'base/login_register.html')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -32,6 +78,7 @@ def room(request, pk):
     context={'room':room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -42,9 +89,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('you shall not PASS!!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -55,8 +106,13 @@ def updateRoom(request, pk):
     context={'form':form}
     return render(request, 'base/room_form.html', context) 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('you shall not PASS!!!')
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
